+++ src/server/game/Handlers/AuctionHouseHandler.cpp
@@ -119,7 +119,7 @@ void WorldSession::SendAuctionOwnerNotification(AuctionEntry* auction)
 //this void creates new auction and adds auction to some auctionhouse
 void WorldSession::HandleAuctionSellItem(WorldPacket& recvData)
 {
-    if (GetPlayer() && (sMailMgr->GetMailBoxSize(GetPlayer()->GetGUID().GetCounter()) + GetPlayer()->GetAuctionLotsCount()) > CONF_GET_INT("Antispam.Mail.Controller"))
+    if (GetPlayer() && (sMailMgr->GetMailBoxSize(GetPlayer()->GetGUID().GetCounter()) + GetPlayer()->GetAuctionLotsCount()) > (uint32)CONF_GET_INT("Antispam.Mail.Controller"))
     {
         GetPlayer()->SendMailResult(0, MAIL_SEND, MAIL_ERR_RECIPIENT_CAP_REACHED);
         recvData.rfinish();
@@ -438,7 +438,7 @@ void WorldSession::HandleAuctionPlaceBid(WorldPacket& recvData)
 {
     LOG_DEBUG("network", "WORLD: Received CMSG_AUCTION_PLACE_BID");
 
-    if (GetPlayer() && (sMailMgr->GetMailBoxSize(GetPlayer()->GetGUID().GetCounter()) + GetPlayer()->GetAuctionLotsCount()) > CONF_GET_INT("Antispam.Mail.Controller"))
+    if (GetPlayer() && (sMailMgr->GetMailBoxSize(GetPlayer()->GetGUID().GetCounter()) + GetPlayer()->GetAuctionLotsCount()) > (uint32)CONF_GET_INT("Antispam.Mail.Controller"))
     {
         GetPlayer()->SendMailResult(0, MAIL_SEND, MAIL_ERR_RECIPIENT_CAP_REACHED);
         recvData.rfinish();
@@ -595,7 +595,7 @@ void WorldSession::HandleAuctionRemoveItem(WorldPacket& recvData)
 {
     LOG_DEBUG("network", "WORLD: Received CMSG_AUCTION_REMOVE_ITEM");
 
-    if (GetPlayer() && (sMailMgr->GetMailBoxSize(GetPlayer()->GetGUID().GetCounter()) + GetPlayer()->GetAuctionLotsCount()) > CONF_GET_INT("Antispam.Mail.Controller"))
+    if (GetPlayer() && (sMailMgr->GetMailBoxSize(GetPlayer()->GetGUID().GetCounter()) + GetPlayer()->GetAuctionLotsCount()) > (uint32)CONF_GET_INT("Antispam.Mail.Controller"))
     {
         GetPlayer()->SendMailResult(0, MAIL_SEND, MAIL_ERR_RECIPIENT_CAP_REACHED);
         recvData.rfinish();
